# docs/openapi.yaml
components:
  schemas:
...
    Artifacts:
      additionalProperties: false
      properties:
        paths:
          items:
            type: string
          type:
            - array
            - "null"
      type: object
    Build:
      additionalProperties: false
      properties:
        args:
          additionalProperties:
            type: string
          type: object
        containerfile:
          type: string
        context:
          type: string
      type: object
...
    EgressRule:
      additionalProperties: false
      properties:
        cidr:
          type: string
        host:
          type: string
      type: object
    Error:
      additionalProperties: false
      properties:
        error:
          $ref: "#/components/schemas/ErrorInfo"
      type: object
    ErrorInfo:
      additionalProperties: false
      properties:
        code:
          description: Stable error code, e.g. not_found, invalid_spec, quota_exceeded.
          type: string
        details:
          description: Every problem, when there are several (invalid_spec, secrets_required, validation).
          items:
            type: string
          type:
            - array
            - "null"
        message:
          type: string
      type: object
...
    Git:
      additionalProperties: false
      properties:
        push:
          $ref: "#/components/schemas/Push"
        repositories:
          items:
            $ref: "#/components/schemas/Repository"
          type:
            - array
            - "null"
      type: object
    History:
      additionalProperties: false
      properties:
        from:
          format: date-time
          type: string
        resolution:
          format: int64
          type: integer
        samples:
          items:
            $ref: "#/components/schemas/Sample"
          type:
            - array
            - "null"
        to:
          format: date-time
          type: string
      type: object
...
    Image:
      additionalProperties: false
      properties:
        build:
          $ref: "#/components/schemas/Build"
        ref:
          type: string
      type: object
    ImageResolution:
      additionalProperties: false
      properties:
        containerfile:
          type: string
        imageId:
          type: string
      type: object
    Init:
      additionalProperties: false
      properties:
        script:
          type: string
      type: object
...
    Network:
      additionalProperties: false
      properties:
        egress:
          items:
            $ref: "#/components/schemas/EgressRule"
          type:
            - array
            - "null"
        ports:
          items:
            $ref: "#/components/schemas/Port"
          type:
            - array
            - "null"
        unrestricted:
          type: boolean
      type: object
...
    Placement:
      additionalProperties: false
      properties:
        pool:
          type: string
        prefers:
          additionalProperties:
            type: string
          type: object
        requires:
          additionalProperties:
            type: string
          type: object
      type: object
...
    Port:
      additionalProperties: false
      properties:
        name:
          type: string
        port:
          format: int64
          type: integer
      type: object
    Push:
      additionalProperties: false
      properties:
        branch:
          type: string
      type: object
...
    Repository:
      additionalProperties: false
      properties:
        credential:
          type: string
        name:
          type: string
        path:
          type: string
        ref:
          type: string
        url:
          type: string
      type: object
...
    Resources:
      additionalProperties: false
      properties:
        cpus:
          format: double
          type: number
        disk:
          description: "A number of bytes, or a size: \"8Gi\", \"512Mi\", \"1G\". Responses carry the number."
          oneOf:
            - format: int64
              type: integer
            - examples:
                - 8Gi
              type: string
        memory:
          description: "A number of bytes, or a size: \"8Gi\", \"512Mi\", \"1G\". Responses carry the number."
          oneOf:
            - format: int64
              type: integer
            - examples:
                - 8Gi
              type: string
        pids:
          format: int64
          type: integer
      type: object
    Resumability:
      additionalProperties: false
      properties:
        blockers:
          items:
            type: string
          type:
            - array
            - "null"
        onHosts:
          items:
            type: string
          type:
            - array
            - "null"
        secrets:
          items:
            type: string
          type:
            - array
            - "null"
        secretsHeld:
          type: boolean
        snapshot:
          type: string
        uploaded:
          type: boolean
      type: object
    Resume:
      additionalProperties: false
      properties:
        command:
          items:
            type: string
          type:
            - array
            - "null"
      type: object
...
    Run:
      additionalProperties: false
      properties:
        activity:
          type: string
        createdAt:
          format: date-time
          type: string
        epoch:
          format: int64
          type: integer
        exitCode:
          format: int64
          type: integer
        finishedAt:
          format: date-time
          type: string
        firstScheduledAt:
          format: date-time
          type: string
        firstStartedAt:
          format: date-time
          type: string
        host:
          description: The name of the host of its current placement.
          type: string
        hostId:
          description: That host's id.
          type: string
        id:
          type: string
        image:
          $ref: "#/components/schemas/ImageResolution"
        labels:
          additionalProperties:
            type: string
          type: object
        name:
          type: string
        placements:
          items:
            $ref: "#/components/schemas/RunPlacement"
          type:
            - array
            - "null"
        resume:
          $ref: "#/components/schemas/Resumability"
        secrets:
          items:
            $ref: "#/components/schemas/SecretRef"
          type:
            - array
            - "null"
        sessionId:
          type: string
        snapshotId:
          type: string
        spec:
          $ref: "#/components/schemas/RunSpec"
        state:
          type: string
        stateReason:
          type: string
        tenant:
          description: The owning tenant's name.
          type: string
        usage:
          $ref: "#/components/schemas/RunUsage"
      type: object
    RunPlacement:
      additionalProperties: false
      properties:
        acceptedAt:
          format: date-time
          type: string
        assignedAt:
          format: date-time
          type: string
        containerStartedAt:
          format: date-time
          type: string
        cpuSeconds:
          format: double
          type: number
        epoch:
          format: int64
          type: integer
        exitCode:
          format: int64
          type: integer
        exitReason:
          type: string
        exitedAt:
          format: date-time
          type: string
        host:
          type: string
        hostName:
          type: string
        imageReadyAt:
          format: date-time
          type: string
        netRxBytes:
          format: int64
          type: integer
        netTxBytes:
          format: int64
          type: integer
        peakDiskBytes:
          format: int64
          type: integer
        peakMemoryBytes:
          format: int64
          type: integer
        peakPids:
          format: int64
          type: integer
        snapshotBytes:
          format: int64
          type: integer
        snapshotDoneAt:
          format: date-time
          type: string
        state:
          type: string
        stopReason:
          type: string
        stopRequestedAt:
          format: date-time
          type: string
        uploadedAt:
          format: date-time
          type: string
        volumesRestoredAt:
          format: date-time
          type: string
        workloadStartedAt:
          format: date-time
          type: string
      type: object
    RunSpec:
      additionalProperties: false
      properties:
        artifacts:
          $ref: "#/components/schemas/Artifacts"
        env:
          additionalProperties:
            type: string
          type: object
        git:
          $ref: "#/components/schemas/Git"
        image:
          $ref: "#/components/schemas/Image"
        init:
          $ref: "#/components/schemas/Init"
        labels:
          additionalProperties:
            type: string
          type: object
        name:
          type: string
        network:
          $ref: "#/components/schemas/Network"
        placement:
          $ref: "#/components/schemas/Placement"
        resources:
          $ref: "#/components/schemas/Resources"
          description: "What the Run gets, and the scheduler reserves on its host. Unset fields take luxd's defaults: cpus 2, memory 8Gi, pids 1024, unless its operator changed them (LUX_DEFAULT_CPUS, LUX_DEFAULT_MEMORY, LUX_DEFAULT_PIDS)."
        sandbox:
          $ref: "#/components/schemas/Sandbox"
        secrets:
          items:
            $ref: "#/components/schemas/Secret"
          type:
            - array
            - "null"
        timeout:
          description: A Go duration ("4h", "90s"), or a number of seconds. Responses carry the string.
          oneOf:
            - examples:
                - 4h
              type: string
            - type: number
        volumes:
          items:
            $ref: "#/components/schemas/Volume"
          type:
            - array
            - "null"
        workload:
          $ref: "#/components/schemas/Workload"
      type: object
    RunUsage:
      additionalProperties: false
      properties:
        cpuSeconds:
          format: double
          type: number
        netRxBytes:
          format: int64
          type: integer
        netTxBytes:
          format: int64
          type: integer
        peakDiskBytes:
          format: int64
          type: integer
        peakMemoryBytes:
          format: int64
          type: integer
        peakPids:
          format: int64
          type: integer
        placements:
          format: int64
          type: integer
        queueSeconds:
          format: double
          type: number
      type: object
    Sample:
      additionalProperties: false
      properties:
        allocCpus:
          format: double
          type: number
        allocMemory:
          format: int64
          type: integer
        allocatedCpus:
          format: double
          type: number
        allocatedMemory:
          format: int64
          type: integer
        at:
          format: date-time
          type: string
        busy:
          format: int64
          type: integer
        capacityCpus:
          format: double
          type: number
        capacityMemory:
          format: int64
          type: integer
        cpuCores:
          format: double
          type: number
        diskBytes:
          format: int64
          type: integer
        epoch:
          format: int64
          type: integer
        finished:
          format: int64
          type: integer
        hosts:
          additionalProperties:
            format: int64
            type: integer
          type: object
        idle:
          format: int64
          type: integer
        memoryBytes:
          format: int64
          type: integer
        netRxRate:
          format: double
          type: number
        netTxRate:
          format: double
          type: number
        pids:
          format: int64
          type: integer
        placements:
          format: int64
          type: integer
        queued:
          format: int64
          type: integer
        runs:
          additionalProperties:
            format: int64
            type: integer
          type: object
        startP50:
          format: double
          type: number
        startP95:
          format: double
          type: number
        started:
          format: int64
          type: integer
      type: object
    Sandbox:
      additionalProperties: false
      properties:
        nestedContainers:
          type: boolean
        readOnlyRoot:
          type: boolean
      type: object
    Secret:
      additionalProperties: false
      properties:
        as:
          type: string
        name:
          type: string
        path:
          type: string
        runnerOnly:
          type: boolean
        value:
          type: string
      type: object
    SecretRef:
      additionalProperties: false
      properties:
        fingerprint:
          type: string
        name:
          type: string
      type: object
...
    Volume:
      additionalProperties: false
      properties:
        kind:
          type: string
        name:
          type: string
        path:
          type: string
      type: object
...
    Workload:
      additionalProperties: false
      properties:
        adapter:
          type: string
        command:
          items:
            type: string
          type:
            - array
            - "null"
        grace:
          description: A Go duration ("4h", "90s"), or a number of seconds. Responses carry the string.
          oneOf:
            - examples:
                - 4h
              type: string
            - type: number
        prompt:
          type: string
        resume:
          $ref: "#/components/schemas/Resume"
        tty:
          type: boolean
        user:
          type: string
        workdir:
          type: string
      type: object
  securitySchemes:
    apiKey:
      description: A tenant API key (luxd admin create-key), or an operator key (luxd admin create-operator-key).
      scheme: bearer
      type: http
info:
  description: |-
    The lux tenant API. Every /v1 call authenticates with an API key (`Authorization: Bearer <key>`); a key has scopes `read`, `run` and `admin`, each implying the ones before it.

    An operator key (`luxd admin create-operator-key`, scope `operator`) belongs to no tenant: it sees and acts on every tenant through the same operations. A Run's operations act in the Run's own tenant; lists span every tenant unless `?tenant=` (an id or a name) narrows them; creating needs `?tenant=`.

    Errors are JSON: `{"error": {"code": "not_found", "message": "...", "details": ["..."]}}`, with a stable `code`.
  title: lux
  version: dev
openapi: 3.1.0
paths:
...
  /v1/artifacts/{aid}:
    get:
      description: |-
        The file the Run wrote, streamed through luxd. Content-Length and X-Lux-SHA256 let a client tell a whole download from one cut short (the response is aborted, never ended cleanly, on a failure mid-way). 409 not_uploaded (with Retry-After) while it is still on its host; 410 gone once deleted by retention.

        Needs the `read` scope.
      operationId: downloadArtifact
      parameters:
        - description: The artifact's id.
          in: path
          name: aid
          required: true
          schema:
            description: The artifact's id.
            type: string
      responses:
        "200":
          content:
            application/octet-stream:
              schema:
                contentMediaType: application/octet-stream
                format: binary
                type: string
          description: The file.
          headers:
            Content-Disposition:
              description: attachment, with the file's name.
              schema:
                type: string
            Content-Length:
              description: The file's size.
              schema:
                type: integer
            X-Lux-SHA256:
              description: The file's SHA-256, hex.
              schema:
                type: string
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "409":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Conflict
        "410":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Gone
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Download an artifact
      tags:
        - runs
...
  /v1/history:
    get:
      description: |-
        Samples of what GET /v1/status reports.

        Needs the `read` scope.
      operationId: history
      parameters:
        - description: "Operator keys: only this tenant (id or name). Tenant keys ignore it."
          explode: false
          in: query
          name: tenant
          schema:
            description: "Operator keys: only this tenant (id or name). Tenant keys ignore it."
            type: string
        - description: "How far back from now: a Go duration or a number of days (7d). Default 1h."
          explode: false
          in: query
          name: since
          schema:
            description: "How far back from now: a Go duration or a number of days (7d). Default 1h."
            examples:
              - 24h
            type: string
        - description: The start (RFC 3339), instead of since.
          explode: false
          in: query
          name: from
          schema:
            description: The start (RFC 3339), instead of since.
            type: string
        - description: The end (RFC 3339); default now.
          explode: false
          in: query
          name: to
          schema:
            description: The end (RFC 3339); default now.
            type: string
        - description: "Resolution in seconds: 0 (raw), 60 or 3600. Default: the finest kept for the range."
          explode: false
          in: query
          name: res
          schema:
            description: "Resolution in seconds: 0 (raw), 60 or 3600. Default: the finest kept for the range."
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/History"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: The state of the system over time
      tags:
        - history
...
  /v1/runs/{id}:
    get:
      description: |-
        With its placements and resource usage.

        Needs the `read` scope.
      operationId: getRun
      parameters:
        - description: The Run's id.
          in: path
          name: id
          required: true
          schema:
            description: The Run's id.
            type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Run"
          description: OK
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unauthorized
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Not Found
        "422":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Unprocessable Entity
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
          description: Internal Server Error
      security:
        - apiKey: []
      summary: Get a Run
      tags:
        - runs
...